fixed the weird int error with floats and then finally an int.......

+++ stepperMotor.py
@@ -105,7 +105,8 @@ class Stepper(object):
   def turn(self,turnInstruction):
     delay = .01
     print "bearing before conversion:", turnInstruction[1]
-    steps = int((int(turnInstruction[1])/360)*512)
+
+    steps = int(float(turnInstruction[1]/360.0)*512.0)
     print "bearing after :", steps
 
     print "In myMotor.turn with turn instructions:", turnInstruction
